+++ src/pamiq_core/interaction/__init__.py
@@ -1,5 +1,5 @@
 from .agent import Agent
-from .environment import Environment
+from .env import Environment
 
 __all__ = [
     "Agent",
