+++ src/pamiq_core/interaction/env.py
