+++ tests/pamiq_core/interaction/test_env.py
